🧪 Refactor test files
