+++ server/test/helpers/connectToDatabase.js
@@ -1,0 +1,18 @@
+const mongoose = require("mongoose");
+
+const connectToDatabase = async () => {
+  if (
+    !process.env.DB_TEST_URL &&
+    !/test-database/.test(process.env.DB_TEST_URL)
+  ) {
+    console.log("Tests can only and must connect to a test database.");
+  }
+
+  const DB = process.env.DB_TEST_URL.replace(
+    "<db_password>",
+    process.env.DB_PASS,
+  );
+  await mongoose.connect(DB);
+};
+
+module.exports = connectToDatabase;
